further refactoring of #13984
