+++ public/app/features/dashboard/dashgrid/PanelHeader/PanelHeaderMenu.tsx
@@ -2,7 +2,7 @@ import React, { PureComponent } from 'react';
 import { DashboardModel } from 'app/features/dashboard/dashboard_model';
 import { PanelModel } from 'app/features/dashboard/panel_model';
 import { PanelHeaderMenuItem } from './PanelHeaderMenuItem';
-import { getPanelMenu } from 'app/features/dashboard/utils/panel_menu';
+import { getPanelMenu } from 'app/features/dashboard/utils/getPanelMenu';
 import { PanelMenuItem } from 'app/types/panel';
 
 export interface Props {
