+++ public/app/features/dashboard/utils/getPanelMenu.ts
